# go/models/model_new_pet.go
/*
 * PetStore
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 2019-12-04T02:15:00Z
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type NewPet struct {

	Type_ *PetType `json:"type,omitempty"`

	Price float32 `json:"price,omitempty"`
}