# go/models/model_new_pet_response.go
/*
 * PetStore
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 2019-12-04T02:15:00Z
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type NewPetResponse struct {

	Pet *Pet `json:"pet,omitempty"`

	Message string `json:"message,omitempty"`
}