# go/models/model_pet_type.go
/*
 * PetStore
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 2019-12-04T02:15:00Z
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type PetType string

// List of PetType
const (
	DOG PetType = "dog"
	CAT PetType = "cat"
	FISH PetType = "fish"
	BIRD PetType = "bird"
	GECKO PetType = "gecko"
)